Metropolis-Hastings Markov-chain Monte Carlo estimation of the b0, b1 and phi parameters from incidence data.

Wrapper function for the fits the b0, b1 and and phi parameters.

Usage

RSVsim_MCMC(
  data,
  cores,
  nchains,
  mcmc_iter,
  mcmc_warm_up,
  prior_functions_list,
  initial_mat,
  fitted_parameter_names,
  fixed_parameters,
  times,
  cohort_step_size,
  init_conds,
  warm_up,
  lower_ll,
  upper_ll
)

Arguments

fitted_parameter_names

Vector of parameter names to exclude from the list because they are fitted. Default: NULL.

times

Simulation times. Default: 0 - 3650 days with intervals of 0.25 days.

cohort_step_size

Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.

init_conds

Initial conditions to run the model. List. Default: NULL. If NULL: 1% RSV prevalence is assumed for people during the primary infection. All other people are assumed to be susceptible to their primary infection.

warm_up

Length of time-points to exclude before calculating the likelihood. Default: NULL.

Value

MCMC samples for b0, b1 and phi.