Get parameters for the RSVsim_log_likelihood function from the data

Helper function to assess the times and ages in the data, and return a list of suitable fixed parameters and the model control option values given as arguments to the function.

Usage

RSVsim_calibration_parameters(
  fitted_parameter_names = c("b0", "b1", "phi"),
  data,
  overrides = list(),
  country = "United Kingdom",
  cohort_step_size = 10,
  warm_up = NULL
)

Arguments

fitted_parameter_names

Vector of parameter names to exclude from the list because they are fitted. Default: NULL.

data

Data frame of incidence data: "incidence", "time" and "age_chr" columns must be included.

overrides

List of default parameters to change.

country

Country for use in the contact_matrix function in the socialmixr package. Can be given as country name or 2 digit ISO code. United Kingdom default.

cohort_step_size

Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.

warm_up

Length of time-points to exclude before calculating the likelihood. Default: NULL.

Value

List of fixed parameters, max_t and warm_up.