Runs the RSV model using odin and dust

Function to run the transmission model with cohort aging.

Usage

RSVsim_run_model_dust(
  parameters,
  times = seq(0, 3650, 0.25),
  cohort_step_size = 10,
  init_conds = NULL,
  warm_up = NULL
)

Arguments

parameters

List of parameters from get_params function.

times

Simulation times. Default: 0 - 3650 days with intervals of 0.25 days.

cohort_step_size

Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.

init_conds

Initial conditions to run the model. List. Default: NULL. If NULL: 1% RSV prevalence is assumed for people during the primary infection. All other people are assumed to be susceptible to their primary infection.

warm_up

Length of time-points to exclude before calculating the likelihood. Default: NULL.

Value

Simulation output (dataframe). In the dataframe, age refers to the lowest age in the age group.