Constrained maximum-likelihood estimation of the b0, b1 and phi parameters from incidence data.

Helper function to assess the times and ages in the data, and return suitable parameters to run the model. Initial parameter values are set to 0.

Usage

constrained_max_likelihood(
  fixed_parameter_list,
  data,
  scale_parameters,
  cohort_step_size,
  dt
)

Arguments

fixed_parameter_list

List of parameters that are fixed.

data

List of data frames of incidence data. For all data frames "incidence", "time" and "age_chr" columns must be included. For incidence data frames a "value" column is needed to contain the incidence values. For the prevalence data frame a "sample" column is required with the number of people tested and a "positive" column with the number of people that test positive.

scale_parameters

List of lower and upper boundaries on the fitted parameters. Default: NULL. If NULL the parameters are not transformed. If a list is provided the parameters log-likelihood is calculated assuming the fitted_parameters have already been scaled between 0 and 1 with the given lower and upper limits.

cohort_step_size

Time steps to run the model over before adjusting the ages of all cohorts. Default: 10 days.

dt

Time steps to run the model over. Default: 0.25 days.

Value

List of fitted parameter values for b0, b1 and phi.