Get parameters for the calibration_likelihood function from the data

Helper function to assess the times and ages in the data, and return suitable parameters to run the model.

Usage

get_calibration_parameters(
  data,
  data_populations,
  country = "United Kingdom",
  warm_up = NULL
)

Arguments

data

List of data frames of incidence data. For all data frames "incidence", "time" and "age_chr" columns must be included. For incidence data frames a "value" column is needed to contain the incidence values. For the prevalence data frame a "sample" column is required with the number of people tested and a "positive" column with the number of people that test positive.

data_populations

Vector of total population sizes for each dataframe in data.

country

Country for use in the contact_matrix function in the socialmixr package. Can be given as country name or 2 digit ISO code. United Kingdom default.

warm_up

Length of time-points to exclude before calculating the likelihood. Default: NULL.

Value

List of fixed parameters, max_t and warm_up.